refactor: ♻️ Extract parse_hostname_for_role function to SSoT utils to be more DRY.
